1. Extracted CustomSlide and Utils; 2. Tweak wordings and code style;

+++ Platform/AptioMemoryFix/ServiceOverrides.c
@@ -291,7 +291,7 @@ MOAllocatePool (
   //
 
   if (Type == EfiBootServicesData) {
-    *Buffer = UmmMalloc (Size);
+    *Buffer = UmmMalloc ((UINT32)Size);
     if (*Buffer)
       return EFI_SUCCESS;
